Move user agent interceptor into a separate module
